+++ core/src/main/scala/geotrellis/server/LayerTms.scala
@@ -44,7 +44,8 @@ object LayerTms {
   def apply[F[_]: Logger: Parallel: Monad, T: TmsReification[F, *]](
     getExpression: F[Expression],
     getParams: F[Map[String, T]],
-    interpreter: Interpreter[F]
+    interpreter: Interpreter[F],
+    cellType: Option[CellType]
   ): (Int, Int, Int) => F[Interpreted[MultibandTile]] = { (z: Int, x: Int, y: Int) =>
     {
       for {
@@ -68,7 +69,7 @@ object LayerTms {
                       case Valid(expression) => interpreter(expression)
                       case Invalid(errors)   => throw new Exception(errors.map(_.repr).reduce)
                     }
-      } yield reified.andThen(_.as[MultibandTile])
+      } yield reified.andThen(_.as[MultibandTile]).andThen(res => Valid(cellType.fold(res)(res.interpretAs)))
     }
   }
 
@@ -79,24 +80,42 @@ object LayerTms {
     mkExpr: Map[String, T] => Expression,
     getParams: F[Map[String, T]],
     interpreter: Interpreter[F]
-  ) = apply[F, T](getParams.map(mkExpr(_)), getParams, interpreter)
+  ): (Int, Int, Int) => F[Interpreted[MultibandTile]] = apply[F, T](getParams.map(mkExpr(_)), getParams, interpreter, None)
 
   /** Provide an expression and expect arguments to fulfill its needs */
   def curried[F[_]: Logger: Parallel: Monad, T: TmsReification[F, *]](
     expr: Expression,
-    interpreter: Interpreter[F]
+    interpreter: Interpreter[F],
+    cellType: Option[CellType]
   ): (Map[String, T], Int, Int, Int) => F[Interpreted[MultibandTile]] =
     (paramMap: Map[String, T], z: Int, x: Int, y: Int) => {
-      val eval = apply[F, T](expr.pure[F], paramMap.pure[F], interpreter)
+      val eval = apply[F, T](expr.pure[F], paramMap.pure[F], interpreter, cellType)
       eval(z, x, y)
     }
 
   /** The identity endpoint (for simple display of raster) */
   def concurrent[F[_]: Logger: Parallel: Monad: Concurrent, T: TmsReification[F, *]](
-    param: T
-  ) =
+    getExpression: F[Expression],
+    getParams: F[Map[String, T]],
+    interpreter: Interpreter[F]
+  ): (Int, Int, Int) => F[Interpreted[MultibandTile]] =
+    apply(getExpression, getParams, interpreter, None)
+
+  /** The identity endpoint (for simple display of raster) */
+  def withCellType[F[_]: Logger: Parallel: Monad: Concurrent, T: TmsReification[F, *]](
+    param: T,
+    cellType: CellType
+  ): (Int, Int, Int) => F[Interpreted[MultibandTile]] =
     (z: Int, x: Int, y: Int) => {
-      val eval = curried[F, T](RasterVar("identity"), ConcurrentInterpreter.DEFAULT)
+      val eval = curried[F, T](RasterVar("identity"), ConcurrentInterpreter.DEFAULT, cellType.some)
+      eval(Map("identity" -> param), z, x, y)
+    }
+
+  def identity[F[_]: Logger: Parallel: Monad: Concurrent, T: TmsReification[F, *]](
+    param: T
+  ): (Int, Int, Int) => F[Interpreted[MultibandTile]] =
+    (z: Int, x: Int, y: Int) => {
+      val eval = curried[F, T](RasterVar("identity"), ConcurrentInterpreter.DEFAULT, None)
       eval(Map("identity" -> param), z, x, y)
     }
 }